Split the front end interface into multiple.
