+++ src/Model/FrontEndAPI/FrontEndExperimenter.cs
@@ -1,7 +1,7 @@
 namespace FrontEndAPI;
 public class FrontEndExperimenter {
 
-    public static void StoreResult(int surveyId, int questionId, int userId, string answer) {
+    public static void StoreResultInDatabase(int surveyId, int questionId, int userId, string answer) {
 
     }
 }
