+++ src/Model/FrontEndAPI/FrontEndSharedSUAndExperimenter.cs
@@ -1,0 +1,8 @@
+namespace FrontEndAPI;
+using Survey;
+
+public static class FrontEndSharedSUAndExperimenter {
+    public static void ImportSurvey(string path) {
+        // Import survey from path
+    }
+}
