+++ src/Model/FrontEndAPI/FrontEndSuperUserMenu.cs
@@ -12,4 +12,8 @@ using Survey;
             return null;
         }
 
+        public static string ExportSurvey(int surveyId) {
+            return null;
+        }
+
     }
